Make DiskLruCache internal, move FileSystem out of internal package, and fix ClassNames
