+++ apollo-compiler/src/main/kotlin/com/apollographql/apollo3/compiler/codegen/ClassNames.kt
@@ -8,7 +8,6 @@ import com.squareup.kotlinpoet.ClassName
 internal object ClassNames {
   const val apolloApiPackageName = "com.apollographql.apollo3.api"
   const val apolloAnnotationsPackageName = "com.apollographql.apollo3.annotations"
-
   const val apolloApiJsonPackageName = "$apolloApiPackageName.json"
   const val apolloApiTestPackageName = "$apolloApiPackageName.test"
