+++ apollo-http-cache/src/main/kotlin/com/apollographql/apollo3/cache/http/FileSystem.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.apollographql.apollo3.cache.http.internal
+package com.apollographql.apollo3.cache.http
 
 import okio.Sink
 import okio.Source
